app/requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
pymysql

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging